Module mysha::ecc

Module for performing elliptic curve cryptography operations.

This module provides tools for dealing with asymetric cryptography over elliptic curves;

It provides types and methods that enable you to use ecc, such as creating private keys, public keys, signing, and performing ecc math, such as point doubling, adding and multiplying.

Examples

Simple generating key-pair and signing:

use mysha::ecc::*;
use mysha::sha256::{sha256, InputType};
use rand::{self, SeedableRng};
use num_bigint::{RandBigInt, BigUint};
 
let curve = Curve::secp256k1();
 
let mut rng = rand::rngs::StdRng::from_entropy(); // check if your rng is cryptographically secure
 
let private_key = rng.gen_biguint_range(&BigUint::from(1_u8), curve.get_n()); // needs to be BigInt, this makes code more complex
 
let key_pair = KeyPair::new(private_key, curve).unwrap();
 
let signature = key_pair.sign("this message needs to be hashed and signed", InputType::Text).unwrap();
 
println!("{:?}", signature); 
 
assert!(signature.verify("this message needs to be hashed and signed", InputType::Text).unwrap());

Structs

Curve

Elliptic Curve type

KeyPair

Key Pair type

PrivKey

Private Key Type

PubKey

Public Key type

Signature

Signature Type

Enums

EccError

The error type implemented for this module, with all possible errors that can occur in ecc operations.

Point

Point type