Struct mysha::ecc::Curve

pub struct Curve { /* private fields */ }

Elliptic Curve type

Contains all the parameters that define an elliptic curve

To create a Curve, refer to new, or to to secp256k1(), to use the secp256k1 standard curve. This methods are necessary to create a Curve, since the fields are private to ensure that only valid elliptic curves are created.

The Curve type is present in most of the other ecc types, as they would be meaningless without its respective originating curve to bound to.

Parameters

the parameters “a” and “b” define the elliptic curve equation, in y² = x³ + ax + b (mod p);

the parameter “p” defines the field of the elliptic curve, so the elliptic curve is (mod p);

the parameter “g” is the curves Generator point, it is a Point that is used to start the ecc operations.

the parameter “n” is the order of the subgroup generated by the curve and the generator point, it is a result of the other parameters, but it needs to be calculated and provided.

Problematic curves

It is worth pointing out, that when creating a curve it is possible to create a problematic curve, despite the verifications made to ensure the curve is valid. This happens because the p and the n parameters might no be prime for the curve in question, and it is infeasable to always check if theese parameters are prime numbers in the process of verifying the curve.

Problematic curves aren’t fit for cryptography, and can cause an NotPrime when doing operations with them, to fix that make sure your curve has prime parameters n and p, and it is a valid curve.

Implementations

impl Curve

pub fn new<T: Into<BigInt> + Into<BigUint>>( a: i32, b: i32, p: T, n: T, g: Point ) -> Result<Curve, EccError>

Creates a new Curve from the curve parameters

It can be called on any type that can be converted into a BigUint, so it needs to be unsigned and an integer. You can also use BigUint itself for bigger numbers.

Examples

let c = Curve::new(2, 3, 97_u32, 50_u32, Point::point(0_u32, 10_u32));

Errors

This can fail if the elliptic curve isn’t valid, or good for cryptography.

pub fn get_a(&self) -> i32

Returns the value of the parameter “a”

pub fn get_b(&self) -> i32

Returns the value of the parameter “b”

pub fn get_p(&self) -> &BigUint

Returns the value of the parameter “p”

pub fn get_n(&self) -> &BigUint

Returns the value of the parameter “n”

pub fn get_g(&self) -> &Point

Returns the generator point

pub fn secp256k1() -> Curve

Returns a Curve with the secp256k1 specs

pub fn is_on_curve(&self, p: &Point) -> bool

Returns a bool value that indicates wether the point provided is on the curve

Examples

let on_curve = c.is_on_curve(&Point::point(10_u32, 76_u32));
assert!(on_curve);

pub fn add(&self, p: &Point, q: &Point) -> Result<Point, EccError>

Adds two points on the Curve

Perform the elliptic curve addition operation on the two points provided.

Examples

let sum = c.add(&Point::point(17_u32, 10_u32), &Point::point(95_u32, 31_u32))?;
assert_eq!(sum, Point::point(1_u32, 54_u32));

Errors

This can fail if the points provided aren’t on the curve, or if there is something wrong with the curve.

pub fn double(&self, p: &Point) -> Result<Point, EccError>

Doubles a Point on the Curve

Performs the elliptic curve double operation on the point provided. Equivalent to adding two equal points.

Examples

let double = c.double(&Point::point(24_u32, 2_u32))?;
assert_eq!(double, Point::point(65_u32, 65_u32));

Errors

This can fail if the point provided isn’t on the curve, or if there is a problem with the curve.

pub fn multiply<T: Into<BigInt>>( &self, p: &Point, k: T ) -> Result<Point, EccError>

Multiples a Point with a scalar number, on the Curve

Performs the multiplication opperation, that consists of multiple add and double operations.

It can be called on any type that can be converted into a BigInt, so it needs to be an integer. You can also use BigInt itself for bigger numbers.

Examples

let n = c.multiply(c.get_g(), 35)?;
assert_eq!(n, Point::point(53_u32, 73_u32));

Errors

This can fail if the Point provided isn’t on the curve, or if there is a problem with the curve.