Struct mysha::sha256::Hash256

pub struct Hash256(/* private fields */);

The return type of the hashing process

To create a Hash256, refer to the from_hex method.

Implementations

impl Hash256

pub fn from_hex(hex: &str, le: bool) -> Result<Hash256, HashError>

Creates a hash type from a hex value.

It can be used with little endian values as well, by setting the le parameter to true.

Examples

 
let hash = Hash256::from_hex("2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824", false)?;
 
assert_eq!(hash, sha256("hello", InputType::Text)?, "Error, hashes don't match");
println!("hashes match!");
 

Errors

If the hash is invalid the function will return a HashError. A hash is invalid if it doesn’t have the correct number of digits, or if the digits aren’t valid as hexadecimal.

 
let x = Hash256::from_hex("abc", false)?;
 

pub fn get_hex(&self) -> &str

Returns the hex digest of the hash.

pub fn get_hex_le(&self) -> String

Returns the hex digest of the hash in little endian byte order.